Feat: 모달 타입 컴포넌트 분리 및 style 파일 분리
