+++ src/components/modals/AddArtistModal.tsx
@@ -1,6 +1,5 @@
 import { useEffect, useRef, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { styled } from 'styled-components';
 
 import closeIcon from '../../assets/icons/close.svg';
 import photoIcon from '../../assets/icons/photo.svg';
@@ -12,46 +11,23 @@ import { useGetArtistsTypeQuery } from '../../redux/artistsTypeSlice';
 import { useAddImageMutation } from '../../redux/imageSlice';
 import { toggleArtist } from '../../redux/manageModalSlice';
 import { artistType } from '../../types/artistsType';
-import SortDropdown from '../SortButton';
 
+import {
+  ModalTitle,
+  ModalCloseButton,
+  TypeTitle,
+  TypeWrapper,
+  ImageNameWrapper,
+  ImagePreview,
+  NameInput,
+  NameLabel,
+  SubmitButton,
+  StyledInput,
+  GroupSortDropdown,
+} from './AddArtistModalStyle';
 import CommonModal from './CommonModal';
 import { ModalPortal } from './CommonModal';
-
-type groupType = {
-  type: string;
-  selected: boolean;
-};
-
-type imageType = {
-  previewimage: string;
-};
-
-type propsType = {
-  data: artistType;
-  selected: boolean;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-};
-
-const ArtistTypeBtn = ({ data, onChange, selected }: propsType) => {
-  return (
-    <>
-      <TypeLabel
-        htmlFor={data.id.toString()}
-        type={data.type}
-        selected={selected}
-      >
-        {data.type}
-      </TypeLabel>
-      <StyledInput
-        type="radio"
-        id={data.id.toString()}
-        name="artistType"
-        value={data.id.toString()}
-        onChange={onChange}
-      />
-    </>
-  );
-};
+import TypeButton from './components/TypeButton';
 
 export type sortOptionsType = {
   sort: string;
@@ -165,7 +141,7 @@ const AddArtistModal = () => {
   let content;
   if (artistTypeArray.length > 0) {
     content = artistTypeArray.map((data) => (
-      <ArtistTypeBtn
+      <TypeButton
         key={data.id}
         data={data}
         onChange={onClickGroupType}
@@ -225,133 +201,3 @@ const AddArtistModal = () => {
 };
 
 export default AddArtistModal;
-
-const ModalTitle = styled.h4`
-  padding: 13px 58px;
-  margin: 24px 0 22px;
-  background-color: #fcf9a4;
-  border-radius: 73px;
-  border: 2.937px solid var(--line-black);
-  font-size: 28px;
-  font-weight: 700;
-  text-align: center;
-`;
-
-const ModalCloseButton = styled.button`
-  position: absolute;
-  top: 10px;
-  right: 17px;
-`;
-
-const TypeTitle = styled.span`
-  font-size: 24px;
-  font-weight: 700;
-`;
-
-const TypeWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-  gap: 15px;
-  width: 70%;
-  margin: 10px 0 32px;
-`;
-
-const TypeLabel = styled.label<groupType>`
-  width: 130px;
-  font-size: 20px;
-  font-weight: 400;
-  text-align: center;
-  padding: 10px 20px;
-  border-radius: 30px;
-  box-shadow: 4.4px 4.4px 0px 0px rgba(0, 0, 0, 0.25);
-  cursor: pointer;
-  ${(props) =>
-    props.selected
-      ? `
-      background-color: #f8f8fa;
-      border: 2px solid var(--line-black);
-     `
-      : `
-      color: #8F9196;
-      border: 2.056px solid #8B8E97;
-      background: #EDEDED;
-      `}
-`;
-
-const ImagePreview = styled.label<imageType>`
-  display: block;
-  position: relative;
-  width: 232px;
-  height: 232px;
-  border: 2px solid var(--line-black);
-  border-radius: 30px;
-  text-align: center;
-  cursor: pointer;
-  ${(props) =>
-    props.previewimage
-      ? `
-        background-image: url(${props.previewimage});
-        background-size: cover;
-        background-position: center center;
-        `
-      : `
-        background-color: var(--line-grey2);
-        `}
-  & > img {
-    position: relative;
-    top: 50%;
-    transform: translateY(-50%);
-  }
-`;
-
-const GroupSortDropdown = styled(SortDropdown)`
-  position: relative;
-  right: 0;
-`;
-const StyledInput = styled.input`
-  position: absolute;
-  clip: rect(0 0 0 0);
-  width: 1px;
-  height: 1px;
-  margin: -1px;
-  overflow: hidden;
-`;
-
-const ImageNameWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 16px;
-  margin-bottom: 26px;
-`;
-
-const NameLabel = styled.label`
-  display: block;
-  font-size: 24px;
-  font-weight: 700;
-  margin: 10px 20px;
-`;
-const NameInput = styled.input`
-  width: 360px;
-  height: 58px;
-  padding: 20px;
-  font-size: 20px;
-  font-weight: 400;
-  background-color: #f8f8fa;
-  border: 1.4px solid var(--font-black);
-  border-radius: 30px;
-  &::placeholder {
-    font-size: 20px;
-    color: 4e5761;
-  }
-`;
-
-const SubmitButton = styled.button`
-  width: 202px;
-  font-size: 35px;
-  font-weight: 700;
-  padding: 16px;
-  border: 3px solid var(--line-black);
-  border-radius: 73px;
-  background-color: #defcf9;
-`;
